feat(dependencyinjection): Create AppModule.kt

+++ app/src/main/java/com/example/weather_app_kotlin_mvvm_repository/data/model/Wind.kt
@@ -7,7 +7,7 @@ data class Wind(
     @SerializedName("deg")
     val deg: Int,
     @SerializedName("gust")
-    val gust: Int,
+    val gust: Double,
     @SerializedName("speed")
     val speed: Double
 )
